requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.34.0
colorclass==2.2.2
colorlog==6.7.0
cron-descriptor==1.4.0
Django==4.2.4
django-celery-beat==2.5.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
et-xmlfile==1.1.0
idna==3.4
install==1.3.5
kombu==5.3.2
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
phonenumbers==8.13.18
Pillow==10.0.0
pip-upgrader==1.4.15
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
terminaltables==3.1.10
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
